Move `Rustup` to the `tools` package
