+++ src/main/kotlin/org/rust/cargo/toolchain/tools/Rustup.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.cargo.toolchain
+package org.rust.cargo.toolchain.tools
 
 import com.intellij.execution.ExecutionException
 import com.intellij.notification.NotificationType
@@ -12,7 +12,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import org.rust.cargo.project.settings.toolchain
-import org.rust.cargo.toolchain.tools.rustc
+import org.rust.cargo.toolchain.RsToolchain
 import org.rust.cargo.util.DownloadResult
 import org.rust.ide.actions.InstallComponentAction
 import org.rust.ide.notifications.showBalloon
